# api/openapi/valori-v1.yaml
components:
  schemas:
    ApiError:
      description: |-
        The canonical error body, as a schema-bearing DTO.

        [`valori_engine::EngineError`] produces this shape at runtime but lives in
        a crate that does not depend on `utoipa`, so the schema is declared here —
        the translation layer §36 asks for, rather than a re-export of an internal
        type.
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable machine-readable code. Branch on this.
        error:
          description: Human-readable message. Do not parse.
          type: string
      required:
      - error
      - code
      type: object
...
    CollectionInfo:
      properties:
        dimension:
          description: Present only for collections created with an explicit vector config.
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
        id:
          format: int32
          minimum: 0
          type: integer
        index:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/IndexKind'
        max_records:
          minimum: 0
          type:
          - integer
          - 'null'
        metric:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Metric'
        name:
          type: string
        record_count:
          minimum: 0
          type:
          - integer
          - 'null'
      required:
      - name
      - id
      type: object
...
    CreateCollectionRequest:
      properties:
        dimension:
          description: |-
            Vector dimension for this collection. **Required for every name** —
            `"default"` included, with no exception (Phase 3.3). Valori does not
            infer a collection's dimension from its first insert or from any
            project/env-level default; `POST` without it is rejected with 400.
            Immutable after creation; a later request with a different value for
            the same collection is rejected, not silently applied.

            The Rust field stays `Option` so that a missing value reaches
            `parse_collection_config` and is answered with that explanatory 400
            rather than a generic deserialization failure. `required = true`
            records the contract-level truth the handler enforces, so a generated
            SDK makes the argument mandatory instead of silently omittable.
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
        index:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/IndexKindInput'
            description: |-
              Index algorithm: `"brute"`, `"hnsw"`, `"ivf"`, `"bq"`, or `"auto"`.
              Optional — omitting it means `index = NONE` (exact
              namespace-specific search, no dedicated ANN structure), a
              first-class supported state, not a missing feature. Requires
              `dimension` to also be set — a collection cannot have an explicit
              index without an explicit dimension, since the index is constructed
              for that dimension at creation time.
        metric:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/MetricInput'
            description: |-
              Distance metric. **Required for every name** — `"default"` included,
              with no exception (Phase 3.3). Sent explicitly over the wire, never
              assumed. Only `"squared_l2"` is supported today; present as a field
              (not hard-coded) so the wire contract does not need to change when a
              second metric is added.

              See `dimension` for why this is `Option` in Rust but `required` in the
              contract.
        name:
          type: string
      required:
      - name
      - dimension
      - metric
      type: object
    CreateCollectionResponse:
      properties:
        created:
          type: boolean
        id:
          format: int32
          minimum: 0
          type: integer
        name:
          type: string
      required:
      - name
      - id
      - created
      type: object
...
    ErrorCode:
      description: |-
        Mirror of [`valori_engine::ErrorCode`] for schema generation.

        `tests/api_contract.rs` diffs the runtime enum against the committed YAML;
        this type exists so the generated document carries the same closed set.
      enum:
      - validation_error
      - unauthorized
      - forbidden
      - not_found
      - collection_not_found
      - record_not_found
      - dimension_mismatch
      - invalid_metric
      - invalid_index
      - index_build_failed
      - conflict
      - capacity_exceeded
      - not_leader
      - unavailable
      - not_implemented
      - internal_error
      type: string
...
    IndexKind:
      description: |-
        The `index` values a response can contain — canonical only, matching
        [`valori_domain::IndexKind`]'s `as_str`.
      enum:
      - brute
      - hnsw
      - ivf
      - bq
      - auto
      type: string
    IndexKindInput:
      description: |-
        The `index` values a request may send — canonical plus the aliases
        [`valori_domain::IndexKind`]'s `FromStr` accepts.

        Omitting `index` entirely is a distinct, first-class state (`index = NONE`,
        exact namespace-scoped search with no dedicated ANN structure) and is not
        represented here — absence is the representation.
      enum:
      - brute
      - bruteforce
      - hnsw
      - ivf
      - bq
      - auto
      - mstg
      type: string
...
    ListCollectionsResponse:
      properties:
        collections:
          items:
            $ref: '#/components/schemas/CollectionInfo'
          type: array
      required:
      - collections
      type: object
...
    Metric:
      description: |-
        The `metric` values a response can contain — canonical only, matching
        [`valori_domain::Metric`]'s `as_str`.
      enum:
      - squared_l2
      type: string
    MetricInput:
      description: |-
        The `metric` values a request may send — canonical plus the aliases
        [`valori_domain::Metric`]'s `FromStr` accepts.
      enum:
      - squared_l2
      - l2
      - l2sq
      type: string
...
    SnapshotSaveRequest:
      properties:
        path:
          type:
          - string
          - 'null'
      type: object
    SnapshotSaveResponse:
      properties:
        path:
          type: string
        success:
          type: boolean
      required:
      - success
      - path
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Project API key or the legacy VALORI_AUTH_TOKEN bearer token.
      scheme: bearer
      type: http
info:
  description: 'The Valori Data Plane REST API v1, emitted directly from `#[utoipa::path]` annotations on the registered axum handlers. Every public route the Rust routers register appears here and nothing else does — `scripts/verify-api-route-contract.py` proves that equality on every run of the contract gate, and fails on any drift in either direction. Administrative and operator-internal routes (key management, cluster membership, replication streams, `/metrics`) are deliberately excluded: they are served by the node but are not part of the SDK surface.'
  license:
    identifier: MIT OR Apache-2.0
    name: MIT OR Apache-2.0
  title: Valori Data Plane API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/namespaces:
    get:
      description: Returns an empty list for a brand-new project. Each entry carries its dimension, metric, index kind, and record count.
      operationId: list_collections
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCollectionsResponse'
          description: Collection inventory
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
      security:
      - BearerAuth: []
      summary: List collections
      tags:
      - collections
      x-required-scope: read_only
      x-sdk: true
    post:
      description: Idempotent. `dimension` and `metric` are always required — a new project has zero collections and `default` carries no implicit config (Phase 3.3).
      operationId: create_collection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCollectionResponse'
          description: Collection created, or already existed with an identical config
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid name, dimension, metric, or index kind
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Collection exists with a conflicting config
        '507':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Record slab capacity exhausted
      security:
      - BearerAuth: []
      summary: Create a collection
      tags:
      - collections
      x-required-scope: read_write
      x-sdk: true
...
  /v1/snapshot/save:
    post:
      description: Writes to `path` when given, otherwise to `VALORI_SNAPSHOT_PATH`. Fails when neither is set.
      operationId: save_snapshot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotSaveRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotSaveResponse'
          description: Snapshot written
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: No snapshot path configured or given
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Write failure
      security:
      - BearerAuth: []
      summary: Write a snapshot to local disk
      tags:
      - snapshot
      x-required-scope: admin
      x-sdk: true
...
servers:
- description: 'The node this document was generated by. Valori is self-hosted: point your client at your own node''s base URL.'
  url: /